backend/pyproject.toml
[project]
name = "rasterviewer"
version = "0.1.0"
description = "GIS TIFF Viewer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "rasterio>=1.3.9",
    "numpy>=1.26.0",
    "Pillow>=10.2.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
]

[tool.ruff]
target-version = "py311"